refactor(apes): consume @openape/shapes; keep grant-orchestration + CLI glue
