+++ packages/shapes/src/audit.ts
