+++ packages/shapes/src/capabilities.ts
