+++ packages/shapes/src/generic.ts
