+++ packages/shapes/test/fixtures/exo.toml
@@ -1,0 +1,44 @@
+schema = "openape-shapes/v1"
+
+[cli]
+id = "exo"
+executable = "exo"
+audience = "shapes"
+version = "1"
+
+[[operation]]
+id = "dns.list"
+command = ["dns", "list"]
+display = "List Exoscale DNS domains"
+action = "list"
+risk = "low"
+resource_chain = ["account:name=current", "dns-domain:*"]
+
+[[operation]]
+id = "dns.show"
+command = ["dns", "show"]
+positionals = ["domain"]
+display = "List DNS records in domain {domain}"
+action = "list"
+risk = "low"
+resource_chain = ["account:name=current", "dns-domain:name={domain}", "dns-record:*"]
+
+[[operation]]
+id = "dns.show.record"
+command = ["dns", "show"]
+positionals = ["domain"]
+required_options = ["name"]
+display = "Read DNS record {name} in domain {domain}"
+action = "read"
+risk = "low"
+resource_chain = ["account:name=current", "dns-domain:name={domain}", "dns-record:name={name}"]
+
+[[operation]]
+id = "dns.remove"
+command = ["dns", "remove"]
+positionals = ["domain", "record"]
+display = "Delete DNS record {record} in domain {domain}"
+action = "delete"
+risk = "high"
+exact_command = true
+resource_chain = ["account:name=current", "dns-domain:name={domain}", "dns-record:name={record}"]
