+++ packages/shapes/test/fixtures/gh.toml
@@ -1,0 +1,255 @@
+schema = "openape-shapes/v1"
+
+[cli]
+id = "gh"
+executable = "gh"
+audience = "shapes"
+version = "1"
+
+# ══════════════════════════════════════════════════════
+# REPO — Read Operations
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "repo.list"
+command = ["repo", "list"]
+positionals = ["owner"]
+display = "List repositories for owner {owner}"
+action = "list"
+risk = "low"
+resource_chain = ["owner:login={owner}", "repo:*"]
+
+[[operation]]
+id = "repo.view"
+command = ["repo", "view"]
+positionals = ["repo"]
+display = "View repository {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}"]
+
+[[operation]]
+id = "repo.clone"
+command = ["repo", "clone"]
+positionals = ["repo"]
+display = "Clone repository {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}"]
+
+# ══════════════════════════════════════════════════════
+# REPO — Write Operations
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "repo.create"
+command = ["repo", "create"]
+positionals = ["name"]
+display = "Create repository {name}"
+action = "create"
+risk = "high"
+exact_command = true
+resource_chain = ["repo:name={name}"]
+
+[[operation]]
+id = "repo.rename"
+command = ["repo", "rename"]
+positionals = ["new_name"]
+required_options = ["repo"]
+display = "Rename repository {repo} to {new_name}"
+action = "rename"
+risk = "high"
+exact_command = true
+resource_chain = ["repo:owner={repo|owner},name={repo|name}"]
+
+[[operation]]
+id = "repo.delete"
+command = ["repo", "delete"]
+positionals = ["repo"]
+display = "Delete repository {repo}"
+action = "delete"
+risk = "critical"
+exact_command = true
+resource_chain = ["repo:owner={repo|owner},name={repo|name}"]
+
+# ══════════════════════════════════════════════════════
+# ISSUES
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "issue.list"
+command = ["issue", "list"]
+required_options = ["repo"]
+display = "List issues in {repo}"
+action = "list"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "issue:*"]
+
+[[operation]]
+id = "issue.view"
+command = ["issue", "view"]
+positionals = ["number"]
+required_options = ["repo"]
+display = "View issue #{number} in {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "issue:number={number}"]
+
+[[operation]]
+id = "issue.create"
+command = ["issue", "create"]
+required_options = ["repo", "title"]
+display = "Create issue in {repo}: {title}"
+action = "create"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "issue:*"]
+
+[[operation]]
+id = "issue.close"
+command = ["issue", "close"]
+positionals = ["number"]
+required_options = ["repo"]
+display = "Close issue #{number} in {repo}"
+action = "edit"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "issue:number={number}"]
+
+[[operation]]
+id = "issue.comment"
+command = ["issue", "comment"]
+positionals = ["number"]
+required_options = ["repo", "body"]
+display = "Comment on issue #{number} in {repo}"
+action = "create"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "issue:number={number}", "comment:*"]
+
+# ══════════════════════════════════════════════════════
+# PULL REQUESTS
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "pr.list"
+command = ["pr", "list"]
+required_options = ["repo"]
+display = "List pull requests in {repo}"
+action = "list"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:*"]
+
+[[operation]]
+id = "pr.view"
+command = ["pr", "view"]
+positionals = ["number"]
+required_options = ["repo"]
+display = "View PR #{number} in {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:number={number}"]
+
+[[operation]]
+id = "pr.create"
+command = ["pr", "create"]
+required_options = ["repo", "title"]
+display = "Create pull request in {repo}: {title}"
+action = "create"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:*"]
+
+[[operation]]
+id = "pr.merge"
+command = ["pr", "merge"]
+positionals = ["number"]
+required_options = ["repo"]
+display = "Merge PR #{number} in {repo}"
+action = "edit"
+risk = "high"
+exact_command = true
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:number={number}"]
+
+[[operation]]
+id = "pr.close"
+command = ["pr", "close"]
+positionals = ["number"]
+required_options = ["repo"]
+display = "Close PR #{number} in {repo}"
+action = "edit"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:number={number}"]
+
+[[operation]]
+id = "pr.comment"
+command = ["pr", "comment"]
+positionals = ["number"]
+required_options = ["repo", "body"]
+display = "Comment on PR #{number} in {repo}"
+action = "create"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "pr:number={number}", "comment:*"]
+
+# ══════════════════════════════════════════════════════
+# RUNS (CI/CD)
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "run.list"
+command = ["run", "list"]
+required_options = ["repo"]
+display = "List workflow runs in {repo}"
+action = "list"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "run:*"]
+
+[[operation]]
+id = "run.view"
+command = ["run", "view"]
+positionals = ["run_id"]
+required_options = ["repo"]
+display = "View run {run_id} in {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "run:id={run_id}"]
+
+[[operation]]
+id = "run.rerun"
+command = ["run", "rerun"]
+positionals = ["run_id"]
+required_options = ["repo"]
+display = "Rerun workflow {run_id} in {repo}"
+action = "edit"
+risk = "medium"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "run:id={run_id}"]
+
+# ══════════════════════════════════════════════════════
+# RELEASE
+# ══════════════════════════════════════════════════════
+
+[[operation]]
+id = "release.list"
+command = ["release", "list"]
+required_options = ["repo"]
+display = "List releases in {repo}"
+action = "list"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "release:*"]
+
+[[operation]]
+id = "release.view"
+command = ["release", "view"]
+positionals = ["tag"]
+required_options = ["repo"]
+display = "View release {tag} in {repo}"
+action = "read"
+risk = "low"
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "release:tag={tag}"]
+
+[[operation]]
+id = "release.create"
+command = ["release", "create"]
+positionals = ["tag"]
+required_options = ["repo"]
+display = "Create release {tag} in {repo}"
+action = "create"
+risk = "high"
+exact_command = true
+resource_chain = ["repo:owner={repo|owner},name={repo|name}", "release:tag={tag}"]
