+++ packages/shapes/test/shapes-adapter.test.ts
@@ -1,19 +1,13 @@
-import type { Server } from 'node:http'
-import { createServer } from 'node:http'
-import { mkdirSync, rmSync, writeFileSync, readFileSync } from 'node:fs'
+import { mkdirSync, readFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import { createPublicKey, generateKeyPairSync, verify } from 'node:crypto'
-import { createRouter, defineEventHandler, readBody, setResponseStatus, toNodeListener } from 'h3'
-import { afterAll, beforeAll, describe, expect, it, vi } from 'vitest'
-import { createIdPApp } from '@openape/server'
-import { SignJWT } from 'jose'
+import { describe, expect, it, vi } from 'vitest'
 
 // ---------------------------------------------------------------------------
 // Isolate HOME to tmpdir
 // ---------------------------------------------------------------------------
 
-const testHome = join(tmpdir(), `apes-shapes-${process.pid}-${Date.now()}`)
+const testHome = join(tmpdir(), `shapes-shapes-${process.pid}-${Date.now()}`)
 mkdirSync(testHome, { recursive: true })
 
 vi.mock('node:os', async (importOriginal) => {
@@ -33,7 +27,7 @@ const FIXTURES_DIR = join(__dirname, 'fixtures')
 
 describe('shapes adapter: TOML parsing', () => {
   it('parses gh.toml into correct adapter structure', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
+    const { parseAdapterToml } = await import('../src/toml')
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
 
@@ -59,7 +53,7 @@ describe('shapes adapter: TOML parsing', () => {
   })
 
   it('parses grep.toml with empty-command operations', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
+    const { parseAdapterToml } = await import('../src/toml')
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
 
@@ -74,17 +68,17 @@ describe('shapes adapter: TOML parsing', () => {
   })
 
   it('rejects invalid TOML (missing schema)', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
+    const { parseAdapterToml } = await import('../src/toml')
     expect(() => parseAdapterToml('[cli]\nid = "x"\nexecutable = "x"\n[[operation]]\nid = "a"\ncommand = []\ndisplay = "d"\naction = "r"\nresource_chain = ["r"]')).toThrow(/schema/)
   })
 
   it('rejects TOML with no operations', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
+    const { parseAdapterToml } = await import('../src/toml')
     expect(() => parseAdapterToml('schema = "openape-shapes/v1"\n[cli]\nid = "x"\nexecutable = "x"')).toThrow(/operation/)
   })
 
   it('parses simple resource chains without commas correctly', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
+    const { parseAdapterToml } = await import('../src/toml')
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
 
@@ -99,8 +93,8 @@ describe('shapes adapter: TOML parsing', () => {
 
 describe('shapes adapter: command resolution', () => {
   it('resolves "gh pr list --repo owner/repo" to pr.list operation', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -124,8 +118,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('resolves "gh repo view owner/repo" with owner/name split', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -139,8 +133,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('resolves "gh repo delete owner/repo" with exact_command constraint', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -154,8 +148,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('resolves "gh repo list owner" with single positional', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -169,8 +163,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('throws for unrecognized command', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -182,8 +176,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('throws when executable does not match adapter', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -195,8 +189,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('resolves grep with positionals', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -213,8 +207,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('resolves grep -r with recursive operation', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -226,8 +220,8 @@ describe('shapes adapter: command resolution', () => {
   })
 
   it('includes execution context with argv hash', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -249,8 +243,8 @@ describe('shapes adapter: command resolution', () => {
 
 describe('shapes adapter: capability resolution', () => {
   it('resolves a capability request for filesystem resource', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -272,8 +266,8 @@ describe('shapes adapter: capability resolution', () => {
   })
 
   it('resolves capability for repo.list (single-segment chain)', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -293,8 +287,8 @@ describe('shapes adapter: capability resolution', () => {
   })
 
   it('throws for unsupported resource chain', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'gh.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -307,8 +301,8 @@ describe('shapes adapter: capability resolution', () => {
   })
 
   it('throws for unsupported action on valid resource', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -321,8 +315,8 @@ describe('shapes adapter: capability resolution', () => {
   })
 
   it('throws when no resources provided', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -335,8 +329,8 @@ describe('shapes adapter: capability resolution', () => {
   })
 
   it('throws when no actions provided', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -355,9 +349,9 @@ describe('shapes adapter: capability resolution', () => {
 
 describe('shapes adapter: grant request building', () => {
   it('builds a structured CLI grant request from resolved command', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
-    const { buildStructuredCliGrantRequest } = await import('../src/shapes/request-builders')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCommand } = await import('../src/parser')
+    const { buildStructuredCliGrantRequest } = await import('../src/request-builders')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -389,7 +383,7 @@ describe('shapes adapter: grant request building', () => {
   })
 
   it('builds an exact command grant request', async () => {
-    const { buildExactCommandGrantRequest } = await import('../src/shapes/request-builders')
+    const { buildExactCommandGrantRequest } = await import('../src/request-builders')
 
     const built = await buildExactCommandGrantRequest(
       ['ls', '-la', '/tmp'],
@@ -410,9 +404,9 @@ describe('shapes adapter: grant request building', () => {
   })
 
   it('builds from capability resolution', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCapabilityRequest } = await import('../src/shapes/capabilities')
-    const { buildStructuredCliGrantRequest } = await import('../src/shapes/request-builders')
+    const { parseAdapterToml } = await import('../src/toml')
+    const { resolveCapabilityRequest } = await import('../src/capabilities')
+    const { buildStructuredCliGrantRequest } = await import('../src/request-builders')
 
     const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
     const adapter = parseAdapterToml(content)
@@ -437,7 +431,7 @@ describe('shapes adapter: grant request building', () => {
   })
 
   it('preserves run_as in request', async () => {
-    const { buildExactCommandGrantRequest } = await import('../src/shapes/request-builders')
+    const { buildExactCommandGrantRequest } = await import('../src/request-builders')
 
     const built = await buildExactCommandGrantRequest(
       ['sudo', 'reboot'],
@@ -452,269 +446,5 @@ describe('shapes adapter: grant request building', () => {
     )
 
     expect(built.request.run_as).toBe('root')
-  })
-})
-
-// ---------------------------------------------------------------------------
-// Test 5: findExistingGrant matching (needs a running IdP)
-// ---------------------------------------------------------------------------
-
-describe('shapes adapter: findExistingGrant', () => {
-  let server: Server
-  let port: number
-  let idpBase: string
-  const MGMT_TOKEN = 'test-mgmt-shapes'
-
-  function generateTestKeyPair() {
-    const { publicKey, privateKey } = generateKeyPairSync('ed25519')
-    const rawPub = publicKey.export({ type: 'spki', format: 'der' })
-    const rawKey = rawPub.subarray(12)
-    const typeStr = 'ssh-ed25519'
-    const typeBuf = Buffer.from(typeStr)
-    const typeLen = Buffer.alloc(4)
-    typeLen.writeUInt32BE(typeBuf.length)
-    const keyLen = Buffer.alloc(4)
-    keyLen.writeUInt32BE(rawKey.length)
-    const wireFormat = Buffer.concat([typeLen, typeBuf, keyLen, rawKey])
-    const publicKeySsh = `ssh-ed25519 ${wireFormat.toString('base64')}`
-    const privateKeyPem = privateKey.export({ type: 'pkcs8', format: 'pem' }) as string
-    return { publicKeySsh, privateKeyPem, privateKey }
-  }
-
-  const agentEmail = 'shapes-agent@example.com'
-  const ownerEmail = 'shapes-owner@example.com'
-  const kp = generateTestKeyPair()
-
-  beforeAll(async () => {
-    // Write agent key
-    writeFileSync(join(testHome, 'test_key'), kp.privateKeyPem, { mode: 0o600 })
-
-    const tempIdp = createIdPApp({ issuer: 'http://placeholder', managementToken: MGMT_TOKEN })
-    const tempServer = createServer(toNodeListener(tempIdp.app))
-    port = await new Promise<number>((resolve, reject) => {
-      tempServer.listen(0, '127.0.0.1', () => {
-        const addr = tempServer.address()
-        if (addr && typeof addr === 'object') resolve(addr.port)
-        else reject(new Error('Failed'))
-      })
-    })
-    await new Promise<void>(resolve => tempServer.close(() => resolve()))
-
-    idpBase = `http://127.0.0.1:${port}`
-    process.env.APES_IDP = idpBase
-
-    const idp = createIdPApp({
-      issuer: idpBase,
-      managementToken: MGMT_TOKEN,
-      adminEmails: [ownerEmail],
-    })
-
-    // Compat routes
-    const { stores } = idp
-    const compatRouter = createRouter()
-
-    compatRouter.post('/api/agent/challenge', defineEventHandler(async (event) => {
-      const body = await readBody<{ agent_id: string }>(event)
-      if (!body.agent_id) { setResponseStatus(event, 400); return { error: 'Missing agent_id' } }
-      const user = await stores.userStore.findByEmail(body.agent_id)
-      if (!user || !user.isActive) { setResponseStatus(event, 404); return { error: 'User not found' } }
-      const challenge = await stores.challengeStore.createChallenge(user.email)
-      return { challenge }
-    }))
-
-    compatRouter.post('/api/agent/authenticate', defineEventHandler(async (event) => {
-      const body = await readBody<{ agent_id: string, challenge: string, signature: string }>(event)
-      if (!body.agent_id || !body.challenge || !body.signature) { setResponseStatus(event, 400); return { error: 'Missing' } }
-      const user = await stores.userStore.findByEmail(body.agent_id)
-      if (!user || !user.isActive) { setResponseStatus(event, 404); return { error: 'Not found' } }
-      const valid = await stores.challengeStore.consumeChallenge(body.challenge, body.agent_id)
-      if (!valid) { setResponseStatus(event, 401); return { error: 'Invalid challenge' } }
-      const keys = await stores.sshKeyStore.findByUser(body.agent_id)
-      if (keys.length === 0) { setResponseStatus(event, 404); return { error: 'No keys' } }
-      let verified = false
-      for (const sshKey of keys) {
-        try {
-          const parts = sshKey.publicKey.trim().split(/\s+/)
-          const keyData = Buffer.from(parts[1]!, 'base64')
-          const tLen = keyData.readUInt32BE(0)
-          const rawKey = keyData.subarray(4 + tLen + 4)
-          const pubKeyObj = createPublicKey({ key: { kty: 'OKP', crv: 'Ed25519', x: rawKey.toString('base64url') }, format: 'jwk' })
-          const sigBuf = Buffer.from(body.signature, 'base64')
-          verified = verify(null, Buffer.from(body.challenge), pubKeyObj, sigBuf)
-          if (verified) break
-        }
-        catch { /* try next */ }
-      }
-      if (!verified) { setResponseStatus(event, 401); return { error: 'Bad sig' } }
-      const signingKey = await stores.keyStore.getSigningKey()
-      const token = await new SignJWT({ sub: user.email, act: user.owner ? 'agent' : 'human' })
-        .setProtectedHeader({ alg: 'EdDSA', kid: signingKey.kid })
-        .setIssuer(idpBase)
-        .setIssuedAt()
-        .setExpirationTime('1h')
-        .sign(signingKey.privateKey)
-      return { token, id: user.email, email: user.email, name: user.name, expires_in: 3600 }
-    }))
-
-    idp.app.use(compatRouter)
-
-    server = createServer(toNodeListener(idp.app))
-    await new Promise<void>((resolve, reject) => {
-      server.listen(port, '127.0.0.1', () => resolve())
-      server.on('error', reject)
-    })
-
-    // Enroll agent
-    const enrollRes = await fetch(`${idpBase}/api/auth/enroll`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({ email: agentEmail, name: 'Shapes Agent', publicKey: kp.publicKeySsh, owner: ownerEmail }),
-    })
-    if (!enrollRes.ok) throw new Error(`Enroll failed: ${await enrollRes.text()}`)
-
-    // Login via apes
-    const { loginCommand } = await import('../src/commands/auth/login')
-    await loginCommand.run!({ args: { idp: idpBase, key: join(testHome, 'test_key'), email: agentEmail } } as any)
-  })
-
-  afterAll(async () => {
-    delete process.env.APES_IDP
-    await new Promise<void>(resolve => server.close(() => resolve()))
-    rmSync(testHome, { recursive: true, force: true })
-  })
-
-  it('finds an existing timed grant that covers the resolved command', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
-    const { findExistingGrant } = await import('../src/shapes/grants')
-
-    const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
-    const adapter = parseAdapterToml(content)
-    const loaded = { adapter, source: join(FIXTURES_DIR, 'grep.toml'), digest: 'test-digest' }
-
-    // Resolve a command to get the authorization detail
-    const resolved = await resolveCommand(loaded, ['grep', 'TODO', '/src'])
-
-    // Create a timed grant via the IdP with matching authorization_details
-    const createRes = await fetch(`${idpBase}/api/grants`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({
-        requester: agentEmail,
-        target_host: 'test-host',
-        audience: 'shapes',
-        grant_type: 'timed',
-        duration: 3600,
-        command: ['grep', 'TODO', '/src'],
-        permissions: [resolved.permission],
-        authorization_details: [resolved.detail],
-        execution_context: resolved.executionContext,
-        reason: 'test timed grant',
-      }),
-    })
-    expect(createRes.status).toBe(201)
-    const created = await createRes.json() as { id: string }
-
-    // Approve it
-    const approveRes = await fetch(`${idpBase}/api/grants/${created.id}/approve`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({ grant_type: 'timed', duration: 3600 }),
-    })
-    expect(approveRes.ok).toBe(true)
-
-    // Now resolve a SIMILAR command (same operation, same path)
-    const resolved2 = await resolveCommand(loaded, ['grep', 'TODO', '/src'])
-
-    // findExistingGrant should find the timed grant
-    const existingId = await findExistingGrant(resolved2, idpBase)
-    expect(existingId).toBe(created.id)
-  })
-
-  it('does not find a once grant for reuse', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
-    const { findExistingGrant } = await import('../src/shapes/grants')
-
-    const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
-    const adapter = parseAdapterToml(content)
-    const loaded = { adapter, source: join(FIXTURES_DIR, 'grep.toml'), digest: 'test-digest' }
-
-    // Create a once grant
-    const resolved = await resolveCommand(loaded, ['grep', '-r', 'FIXME', '/app'])
-
-    const createRes = await fetch(`${idpBase}/api/grants`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({
-        requester: agentEmail,
-        target_host: 'test-host',
-        audience: 'shapes',
-        grant_type: 'once',
-        command: ['grep', '-r', 'FIXME', '/app'],
-        permissions: [resolved.permission],
-        authorization_details: [resolved.detail],
-        reason: 'once grant test',
-      }),
-    })
-    expect(createRes.status).toBe(201)
-    const created = await createRes.json() as { id: string }
-
-    // Approve it
-    await fetch(`${idpBase}/api/grants/${created.id}/approve`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({}),
-    })
-
-    // findExistingGrant should NOT return this once-grant
-    const resolved2 = await resolveCommand(loaded, ['grep', '-r', 'FIXME', '/app'])
-    const existingId = await findExistingGrant(resolved2, idpBase)
-    // Should not match the once grant
-    if (existingId) {
-      expect(existingId).not.toBe(created.id)
-    }
-  })
-
-  it('does not find a grant with mismatched audience', async () => {
-    const { parseAdapterToml } = await import('../src/shapes/toml')
-    const { resolveCommand } = await import('../src/shapes/parser')
-    const { findExistingGrant } = await import('../src/shapes/grants')
-
-    const content = readFileSync(join(FIXTURES_DIR, 'grep.toml'), 'utf-8')
-    const adapter = parseAdapterToml(content)
-
-    // Create a timed grant with a DIFFERENT audience
-    const createRes = await fetch(`${idpBase}/api/grants`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({
-        requester: agentEmail,
-        target_host: 'test-host',
-        audience: 'wrong-audience',
-        grant_type: 'timed',
-        duration: 3600,
-        command: ['grep', 'test', '/etc'],
-        reason: 'wrong audience test',
-      }),
-    })
-    expect(createRes.status).toBe(201)
-    const created = await createRes.json() as { id: string }
-
-    await fetch(`${idpBase}/api/grants/${created.id}/approve`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${MGMT_TOKEN}` },
-      body: JSON.stringify({ grant_type: 'timed', duration: 3600 }),
-    })
-
-    // Use the grep adapter which expects audience "shapes"
-    const loaded = { adapter, source: join(FIXTURES_DIR, 'grep.toml'), digest: 'test-digest' }
-    const resolved = await resolveCommand(loaded, ['grep', 'test', '/etc'])
-    const existingId = await findExistingGrant(resolved, idpBase)
-
-    // Should not match the wrong-audience grant
-    if (existingId) {
-      expect(existingId).not.toBe(created.id)
-    }
   })
 })
